Rename views-dsl-design to views-dsl-material
