+++ sample/src/main/java/com/louiscad/splittiessample/about/AboutUi.kt
@@ -26,8 +26,8 @@ import splitties.views.dsl.core.Ui
 import splitties.views.dsl.core.add
 import splitties.views.dsl.core.textView
 import splitties.views.dsl.core.wrapContent
-import splitties.views.dsl.design.contentScrollingWithAppBarLParams
-import splitties.views.dsl.design.coordinatorLayout
+import splitties.views.dsl.material.contentScrollingWithAppBarLParams
+import splitties.views.dsl.material.coordinatorLayout
 import splitties.views.textAppearance
 import splitties.views.textResource
