+++ settings.gradle.kts
@@ -50,8 +50,8 @@ include(
         ":views-dsl",
         ":views-dsl-appcompat",
         ":views-dsl-constraintlayout",
-        ":views-dsl-design",
         ":views-dsl-ide-preview",
+        ":views-dsl-material",
         ":views-dsl-recyclerview",
         ":views-material",
         ":views-recyclerview",
