+++ views-dsl-material/src/main/java/splitties/views/dsl/material/AppBarLayout.kt
@@ -16,7 +16,7 @@
 
 @file:Suppress("NOTHING_TO_INLINE")
 
-package splitties.views.dsl.design
+package splitties.views.dsl.material
 
 import com.google.android.material.appbar.AppBarLayout
 import splitties.views.dsl.core.matchParent
